fix: move node tests to node folder
