+++ packages/ipfs-http-client/test/node.js
@@ -1,2 +1,4 @@
 'use strict'
 require('./node/swarm')
+require('./node/request-api')
+require('./node/custom-headers')
