+++ packages/ipfs-http-client/test/node/custom-headers.js
@@ -1,45 +1,101 @@
 /* eslint-env mocha */
 'use strict'
 
+const { isNode } = require('ipfs-utils/src/env')
 const { expect } = require('interface-ipfs-core/src/utils/mocha')
 const ipfsClient = require('../../src')
 
-describe('custom headers', function () {
-  let ipfs
-  // initialize ipfs with custom headers
-  before(() => {
-    ipfs = ipfsClient({
-      host: 'localhost',
-      port: 6001,
-      protocol: 'http',
-      headers: {
-        authorization: 'Bearer ' + 'YOLO'
-      }
+function startServer (fn) {
+  let headersResolve
+  const headers = new Promise((resolve) => {
+    headersResolve = resolve
+  })
+
+  // spin up a test http server to inspect the requests made by the library
+  const server = require('http').createServer((req, res) => {
+    req.on('data', () => {})
+    req.on('end', () => {
+      res.writeHead(200)
+      res.write(JSON.stringify({}))
+      res.end()
+      server.close()
+
+      headersResolve(req.headers)
     })
   })
 
-  it('are supported', (done) => {
-    // spin up a test http server to inspect the requests made by the library
-    const server = require('http').createServer((req, res) => {
-      req.on('data', () => {})
-      req.on('end', () => {
-        res.writeHead(200)
-        res.write(JSON.stringify({}))
-        res.end()
-        // ensure custom headers are present
-        expect(req.headers.authorization).to.equal('Bearer ' + 'YOLO')
-        server.close()
-        done()
+  server.listen(6001, () => {
+    fn().then(() => {}, () => {})
+  })
+
+  return headers
+}
+
+describe('custom headers', function () {
+  // do not test in browser
+  if (!isNode) {
+    return
+  }
+
+  let ipfs
+
+  describe('supported in the constructor', () => {
+    // initialize ipfs with custom headers
+    before(() => {
+      ipfs = ipfsClient({
+        host: 'localhost',
+        port: 6001,
+        protocol: 'http',
+        headers: {
+          authorization: 'Bearer YOLO'
+        }
       })
     })
 
-    server.listen(6001, () => {
-      ipfs.id((err, res) => {
-        if (err) {
-          throw err
-        }
-        // this call is used to test that headers are being sent.
+    it('regular API calls', async () => {
+      const headers = await startServer(() => ipfs.id())
+
+      expect(headers.authorization).to.equal('Bearer YOLO')
+    })
+
+    it('multipart API calls', async () => {
+      const headers = await startServer(() => ipfs.files.write('/foo/bar', Buffer.from('derp'), {
+        create: true
+      }))
+
+      expect(headers.authorization).to.equal('Bearer YOLO')
+    })
+  })
+
+  describe('supported as API call arguemnts', () => {
+    // initialize ipfs with custom headers
+    before(() => {
+      ipfs = ipfsClient({
+        host: 'localhost',
+        port: 6001,
+        protocol: 'http'
       })
+    })
+
+    it('regular API calls', async () => {
+      const headers = await startServer(() => ipfs.id({
+        headers: {
+          authorization: 'Bearer OLOY'
+        }
+      }))
+
+      expect(headers.authorization).to.equal('Bearer OLOY')
+    })
+
+    it('multipart API calls', async () => {
+      const headers = await startServer(() => ipfs.files.write('/foo/bar', Buffer.from('derp'), {
+        create: true,
+        headers: {
+          authorization: 'Bearer OLOY'
+        }
+      }))
+
+      expect(headers.authorization).to.equal('Bearer OLOY')
     })
   })
 })
